added to create* functions, need to validate security on functions, need to figure out rewards and disputes, need to add rewards to post_txn route

+++ src/xref/state/txn.rs
@@ -7,7 +7,7 @@ pub struct TxnData {
     // ad_pda_addr: &Pubkey,
     pub merchant_signature: Vec<u8>,
     pub user_signature: Vec<u8>,
-    pub cost: u8,// this is in sol
+    pub cost: u64,// this is in sol
     // the bottom 2 are constructed locally and should not be passed in by sender
     //pub global_txn_id: Vec<u8>, // this is the hash of the user/merchant/referrer
     //pub local_txn_id: Vec<u8>, // this is the hash of the user/merchant
